Separando logica do controller para listagem das tabelas no frontend.
