+++ src/controller/DiscenteController.js
@@ -1,4 +1,3 @@
-const connection = require('../database/index');
 const Discente = require('../models/Discente');
 
 module.exports = {
@@ -75,21 +74,4 @@ module.exports = {
     });
   },
 
-  async list(req, res) {
-    const { table_name } = req.headers;
-    const { query_front } = req.body;
-
-    if (table_name) {
-      const query = `SELECT * FROM \"${table_name}\" LIMIT 40`
-      const [result] = await connection.query(query);
-      return res.json(result);
-    }
-
-    if (query_front) {
-      const [result] = await connection.query(query_front);
-      return res.json(result);
-    }
-
-    return res.status(400).json({ error: 'header and/or body not sent' })
-  }
 }
